Move adapter support modules into their new home.
