+++ lib/data_mapper/constraints/adapters/abstract_adapter.rb
