+++ lib/data_mapper/constraints/adapters/do_adapter.rb
