+++ lib/data_mapper/constraints/adapters/postgres_adapter.rb
@@ -1,10 +1,10 @@
-require 'dm-constraints/adapters/dm-do-adapter'
+require 'data_mapper/constraints/adapters/do_adapter'
 
 module DataMapper
   module Constraints
     module Adapters
 
-      module SqlserverAdapter
+      module PostgresAdapter
         include DataObjectsAdapter
       end
