+++ lib/data_mapper/constraints/adapters/sqlite_adapter.rb
